requirements.txt
opencv-python
numpy
onnxruntime
requests
python-dotenv
pyyaml